remove bgpattributesroute
