+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/BgpRoute.java
@@ -1,5 +1,6 @@
 package org.batfish.datamodel;
 
+import static com.google.common.base.MoreObjects.firstNonNull;
 import static com.google.common.base.Preconditions.checkArgument;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
@@ -9,6 +10,7 @@ import com.google.common.cache.CacheLoader;
 import com.google.common.cache.LoadingCache;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.ImmutableSortedSet;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.Set;
@@ -18,12 +20,18 @@ import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
 import org.batfish.datamodel.BgpRoute.Builder;
 import org.batfish.datamodel.bgp.community.Community;
+import org.batfish.datamodel.bgp.community.ExtendedCommunity;
+import org.batfish.datamodel.bgp.community.StandardCommunity;
 import org.batfish.datamodel.route.nh.NextHop;
+import org.batfish.datamodel.routing_policy.communities.CommunitySet;
 
 /** A generic BGP route containing the common properties among different types of BGP routes */
 @ParametersAreNonnullByDefault
 public abstract class BgpRoute<B extends Builder<B, R>, R extends BgpRoute<B, R>>
-    extends BgpAttributesRoute<B, R> {
+    extends AbstractRoute {
+
+  /** Local-preference has a maximum value of u32 max. */
+  public static final long MAX_LOCAL_PREFERENCE = (1L << 32) - 1;
 
   // Soft values: let it be garbage collected in times of pressure.
   // Maximum size 2^16: Just some upper bound on cache size, well less than GiB.
@@ -37,18 +45,50 @@ public abstract class BgpRoute<B extends Builder<B, R>, R extends BgpRoute<B, R>
   /** Builder for {@link BgpRoute} */
   @ParametersAreNonnullByDefault
   public abstract static class Builder<B extends Builder<B, R>, R extends BgpRoute<B, R>>
-      extends BgpAttributesRoute.Builder<B, R> {
+      extends AbstractRouteBuilder<B, R> {
 
+    @Nonnull protected AsPath _asPath;
+    // Invariant: either immutable or a local copy shielded from external mutations.
     @Nonnull protected Set<Long> _clusterList;
     // Invariant: either immutable or a local copy shielded from external mutations.
+    @Nonnull protected Set<Community> _communities;
+    protected long _localPreference;
     @Nullable protected Ip _originatorIp;
+    @Nullable protected OriginType _originType;
     @Nullable protected RoutingProtocol _protocol;
     @Nullable protected Ip _receivedFromIp;
     protected boolean _receivedFromRouteReflectorClient;
     @Nullable protected RoutingProtocol _srcProtocol;
+    protected int _weight;
 
     protected Builder() {
+      _asPath = AsPath.empty();
+      _communities = ImmutableSet.of();
       _clusterList = ImmutableSet.of();
+    }
+
+    /**
+     * Returns a completely new builder of type {@link B} which has all the fields unset.
+     *
+     * @return A completely new builder of type {@link B}.
+     */
+    /* This is needed in cases where we need to create a new builder having type same as any of the
+    subclasses of BgpRoute's builder but we are not sure of the exact type of the concrete child
+    class.
+    For example while evaluating a routing policy and executing its statements we need
+    to create a completely new builder which should be of the same type as environment's output
+    route builder but we are not sure of the concrete type and only know that it extends the
+    abstract BgpRoute's builder. */
+    @Nonnull
+    public abstract B newBuilder();
+
+    @Nonnull
+    @Override
+    public abstract R build();
+
+    @Nonnull
+    public AsPath getAsPath() {
+      return _asPath;
     }
 
     @Nonnull
@@ -58,14 +98,43 @@ public abstract class BgpRoute<B extends Builder<B, R>, R extends BgpRoute<B, R>
           : Collections.unmodifiableSet(_clusterList);
     }
 
+    @Nonnull
+    public Set<Community> getCommunities() {
+      return _communities instanceof ImmutableSet
+          ? _communities
+          : Collections.unmodifiableSet(_communities);
+    }
+
+    public long getLocalPreference() {
+      return _localPreference;
+    }
+
     @Nullable
     public Ip getOriginatorIp() {
       return _originatorIp;
     }
 
     @Nullable
+    public OriginType getOriginType() {
+      return _originType;
+    }
+
+    @Nullable
     public RoutingProtocol getProtocol() {
       return _protocol;
+    }
+
+    @Override
+    @Nonnull
+    protected abstract B getThis();
+
+    public int getWeight() {
+      return _weight;
+    }
+
+    public B setAsPath(AsPath asPath) {
+      _asPath = asPath;
+      return getThis();
     }
 
     /** Overwrite the clusterList attribute */
@@ -92,8 +161,63 @@ public abstract class BgpRoute<B extends Builder<B, R>, R extends BgpRoute<B, R>
       return getThis();
     }
 
+    /** Overwrite communities */
+    public B setCommunities(CommunitySet communities) {
+      _communities = communities.getCommunities();
+      return getThis();
+    }
+
+    /** Overwrite communities */
+    public B setCommunities(Collection<? extends Community> communities) {
+      if (communities instanceof ImmutableSet) {
+        @SuppressWarnings("unchecked") // cannot be mutated, cast to superclass is safe.
+        ImmutableSet<Community> immutableComm = (ImmutableSet<Community>) communities;
+        _communities = immutableComm;
+      } else {
+        _communities = new HashSet<>(communities);
+      }
+      return getThis();
+    }
+
+    /** Add communities */
+    public B addCommunities(Collection<? extends Community> communities) {
+      if (_communities instanceof ImmutableSet) {
+        _communities = new HashSet<>(_communities);
+      }
+      _communities.addAll(communities);
+      return getThis();
+    }
+
+    /** Add a single community */
+    public B addCommunity(Community community) {
+      if (_communities instanceof ImmutableSet) {
+        _communities = new HashSet<>(_communities);
+      }
+      _communities.add(community);
+      return getThis();
+    }
+
+    /** Add communities */
+    public B removeCommunities(Set<Community> communities) {
+      if (_communities instanceof ImmutableSet) {
+        _communities = new HashSet<>(_communities);
+      }
+      _communities.removeAll(communities);
+      return getThis();
+    }
+
+    public B setLocalPreference(long localPreference) {
+      _localPreference = localPreference;
+      return getThis();
+    }
+
     public B setOriginatorIp(Ip originatorIp) {
       _originatorIp = originatorIp;
+      return getThis();
+    }
+
+    public B setOriginType(OriginType originType) {
+      _originType = originType;
       return getThis();
     }
 
@@ -116,17 +240,35 @@ public abstract class BgpRoute<B extends Builder<B, R>, R extends BgpRoute<B, R>
       _srcProtocol = srcProtocol;
       return getThis();
     }
+
+    public B setWeight(int weight) {
+      _weight = weight;
+      return getThis();
+    }
   }
 
+  /** Default local preference for a BGP route if one is not set explicitly */
+  public static final long DEFAULT_LOCAL_PREFERENCE = 100L;
+
+  public static final String PROP_AS_PATH = "asPath";
   static final String PROP_CLUSTER_LIST = "clusterList";
+  public static final String PROP_COMMUNITIES = "communities";
+  public static final String PROP_LOCAL_PREFERENCE = "localPreference";
+  static final String PROP_ORIGIN_TYPE = "originType";
   static final String PROP_ORIGINATOR_IP = "originatorIp";
   static final String PROP_RECEIVED_FROM_IP = "receivedFromIp";
   static final String PROP_RECEIVED_FROM_ROUTE_REFLECTOR_CLIENT =
       "receivedFromRouteReflectorClient";
   static final String PROP_SRC_PROTOCOL = "srcProtocol";
+  static final String PROP_WEIGHT = "weight";
 
+  @Nonnull protected final AsPath _asPath;
   @Nonnull protected final Set<Long> _clusterList;
+  @Nonnull protected final CommunitySet _communities;
+  protected final long _localPreference;
+  protected final long _med;
   @Nonnull protected final Ip _originatorIp;
+  @Nonnull protected final OriginType _originType;
   @Nonnull protected final RoutingProtocol _protocol;
 
   /**
@@ -139,6 +281,8 @@ public abstract class BgpRoute<B extends Builder<B, R>, R extends BgpRoute<B, R>
 
   protected final boolean _receivedFromRouteReflectorClient;
   @Nullable protected final RoutingProtocol _srcProtocol;
+  /* NOTE: Cisco-only attribute */
+  protected final int _weight;
 
   protected BgpRoute(
       @Nullable Prefix network,
@@ -159,42 +303,79 @@ public abstract class BgpRoute<B extends Builder<B, R>, R extends BgpRoute<B, R>
       int weight,
       boolean nonForwarding,
       boolean nonRouting) {
-    super(
-        network,
-        nextHop,
-        admin,
-        asPath,
-        communities,
-        localPreference,
-        med,
-        originType,
-        tag,
-        weight,
-        nonForwarding,
-        nonRouting);
+    super(network, admin, tag, nonRouting, nonForwarding);
     checkArgument(
         protocol == RoutingProtocol.BGP
             || protocol == RoutingProtocol.IBGP
             || protocol == RoutingProtocol.AGGREGATE,
         "Invalid BgpRoute protocol");
+    _asPath = firstNonNull(asPath, AsPath.empty());
     _clusterList =
         clusterList == null ? ImmutableSet.of() : CLUSTER_CACHE.getUnchecked(clusterList);
+    _communities = communities == null ? CommunitySet.empty() : CommunitySet.of(communities);
+    _localPreference = localPreference;
+    _med = med;
     _nextHop = nextHop;
     _originatorIp = originatorIp;
+    _originType = originType;
     _protocol = protocol;
     _receivedFromIp = receivedFromIp;
     _receivedFromRouteReflectorClient = receivedFromRouteReflectorClient;
     _srcProtocol = srcProtocol;
+    _weight = weight;
+  }
+
+  @Nonnull
+  @JsonProperty(PROP_AS_PATH)
+  public AsPath getAsPath() {
+    return _asPath;
   }
 
   public @Nonnull Set<Long> getClusterList() {
     return _clusterList;
   }
 
+  /** Return the set of all community attributes */
+  public @Nonnull CommunitySet getCommunities() {
+    return _communities;
+  }
+
+  /** Return only standard community attributes */
+  @Nonnull
+  @JsonIgnore
+  public Set<StandardCommunity> getStandardCommunities() {
+    return _communities.getStandardCommunities();
+  }
+
+  /** Return only extended community attributes */
+  @Nonnull
+  @JsonIgnore
+  public Set<ExtendedCommunity> getExtendedCommunities() {
+    return _communities.getExtendedCommunities();
+  }
+
+  @JsonProperty(PROP_LOCAL_PREFERENCE)
+  public long getLocalPreference() {
+    return _localPreference;
+  }
+
+  @JsonIgnore(false)
+  @JsonProperty(PROP_METRIC)
+  @Override
+  public Long getMetric() {
+    return _med;
+  }
+
   @Nonnull
   @JsonProperty(PROP_ORIGINATOR_IP)
   public Ip getOriginatorIp() {
     return _originatorIp;
+  }
+
+  @Nonnull
+  @JsonProperty(PROP_ORIGIN_TYPE)
+  public OriginType getOriginType() {
+    return _originType;
   }
 
   @Nonnull
@@ -222,8 +403,21 @@ public abstract class BgpRoute<B extends Builder<B, R>, R extends BgpRoute<B, R>
     return _srcProtocol;
   }
 
+  @JsonProperty(PROP_WEIGHT)
+  public int getWeight() {
+    return _weight;
+  }
+
+  @Override
+  public abstract B toBuilder();
+
   @JsonProperty(PROP_CLUSTER_LIST)
   private @Nonnull SortedSet<Long> getJsonClusterList() {
     return ImmutableSortedSet.copyOf(_clusterList);
+  }
+
+  @JsonProperty(PROP_COMMUNITIES)
+  private @Nonnull CommunitySet getJsonCommunities() {
+    return _communities;
   }
 }
