+++ projects/batfish-common-protocol/src/test/java/org/batfish/datamodel/routing_policy/expr/DecrementLocalPreferenceTest.java
@@ -1,6 +1,6 @@
 package org.batfish.datamodel.routing_policy.expr;
 
-import static org.batfish.datamodel.BgpAttributesRoute.MAX_LOCAL_PREFERENCE;
+import static org.batfish.datamodel.BgpRoute.MAX_LOCAL_PREFERENCE;
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.Assert.assertThat;
